+++ xrGame/game_cl_base.h
@@ -99,7 +99,7 @@ public:
 	game_PlayerState *GetPlayerByGameID(u32 GameID);
 	game_PlayerState *GetPlayerByOrderID(u32 id);
 	ClientID GetClientIDByOrderID(u32 id);
-	u32 GetPlayersCount() const { return players.size(); };
+	u32 GetPlayersCount() const { return static_cast<u32>(players.size()); };
 	virtual CUIGameCustom *createGameUI() { return NULL; };
 
 	void ReceivePatrolPaths(NET_Packet& P);
