+++ xrGame/ui/UIListWnd.h
@@ -65,7 +65,7 @@ public:
 	void SetAlwaysShowScroll(bool flag = true) { m_bAlwaysShowScroll = flag; }
 	void EnableAlwaysShowScroll(bool flag) { m_bAlwaysShowScroll_enable = flag; }
 
-	int GetItemsCount() { return m_ItemList.size(); }
+	int GetItemsCount() { return static_cast<int>(m_ItemList.size()); }
 
 	//подготовить все элементы заново
 	void Reset();
