+++ xrGame/ui/UIWindow.h
@@ -34,7 +34,7 @@ public:
 	virtual void DetachChild(CUIWindow *pChild);
 	virtual bool IsChild(CUIWindow *pChild) const;
 	virtual void DetachAll();
-	int GetChildNum() { return m_ChildWndList.size(); }
+	int GetChildNum() { return static_cast<int>(m_ChildWndList.size()); }
 
 	void SetParent(CUIWindow *pNewParent);
 	CUIWindow *GetParent() const { return m_pParentWnd; }
